+++ MWDPlanner/SharedObjects/TaskTableView.swift
@@ -4,17 +4,28 @@ import UIKit
 // UITableView의 display와 manage를 책임지는 클래스
 class TaskTableView: NSObject {
     
+    private var tasks: [Task] = []
+    private let taskService: TaskService
+    
     let tableView: UITableView = {
         let tableView = UITableView()
         tableView.register(UITableViewCell.self, forCellReuseIdentifier: "TaskCell")
         return tableView
     }()
+
     
-    let taskTableViewModel: TaskTableViewModel
-    
-    init(taskTableViewModel: TaskTableViewModel) {
-        self.taskTableViewModel = taskTableViewModel
+     init(service: TaskService) {
+         self.taskService = service
+       
+
         super.init()
+         
+         // 클로저를 통해 모델과 연결
+         self.taskService.fetchCompletion = { [weak self] taskArray in
+             self?.tasks = taskArray
+         }
+         tasks = taskService.fetchTasks()
+
         tableView.dataSource = self
         tableView.delegate = self
     }
@@ -24,13 +35,14 @@ class TaskTableView: NSObject {
 extension TaskTableView: UITableViewDataSource, UITableViewDelegate {
     func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
         
-        taskTableViewModel.numberOfTasks
+        self.tasks.count
     }
 
     func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
         let cell = tableView.dequeueReusableCell(withIdentifier: "TaskCell", for: indexPath)
-        let task = taskTableViewModel.task(at: indexPath.row)
-        print(task.title)
+      
+        let task = tasks[indexPath.row]
+        print(task)
         cell.textLabel?.text = task.title
         return cell
     }
